% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{changeText}
\alias{changeText}
\title{Ersetzt Text in Textelement}
\usage{
changeText(svg, element_name, text, alignment = NULL, in_group = NULL,
  hide_blank = FALSE)
}
\arguments{
\item{svg}{SVG als XML document}

\item{element_name}{Name des Textelements bzw. Textinhalt des Textelements, wenn dieses nicht gesondert beschriftet ist.}

\item{text}{Text der eingetragen werden soll.}

\item{alignment}{Textausrichtung. Moegliche Parameter: start, middle, end. (Default NULL)}

\item{in_group}{In welcher Gruppe befindet sich das Textelement (Default NULL).}

\item{hide_blank}{Soll Textelement versteckt werden, wenn es leer ist? (Default FALSE)}
}
\value{
adaptiertes SVG als XML document
}
\description{
Passt ein in der SVG-Datei vorhandenes Textelement an. Kann entweder via Textinhalt selbst oder via Benennung des Textelements angewendet werden.
}
