% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{diffBar}
\alias{diffBar}
\title{Passt Differenzbalkendiagramm an}
\usage{
diffBar(svg, frame_name, frame0_name, group_name, scale_real, values,
  alignment = "horizontal", has_labels = TRUE,
  label_position = "center", decimals = 0, display_limits = c(0, 0),
  label_edge = 10, ignore_rotation = TRUE)
}
\arguments{
\item{svg}{SVG als XML document}

\item{frame_name}{Name des Rechtseck-Elements mit dem Rahmen des Diagrammbereichs}

\item{frame0_name}{Name der 0er-Linie}

\item{group_name}{Name der Gruppe mit den Balkenelementen bzw. mehreren Balken}

\item{scale_real}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100))}

\item{values}{Dataframe mit den Werten, eine Zeile pro Balken}

\item{alignment}{Ausrichtung der Balken. Entweder "horizontal" (default) oder "vertical"}

\item{has_labels}{Sollen Wertelabels angezeigt werden? (Default TRUE)}

\item{label_position}{Position der Wertelabels (Default "center")}

\item{decimals}{Anzahl der Dezimalstellen der Wertelabels (Default 0)}

\item{display_limits}{Zwischen welchen Werten (Minimum und Maximum) rund um Null sollen Wertelabels unterdrückt werden? (Default c(0,0) = keine unterdrückten Werte)}

\item{label_edge}{Abstand der Labels zum Balkenmaximum bei label_position start/end. (Default 10)}

\item{ignore_rotation}{Soll ggf. Rotation der vorbereiteten Textelemente ignoriert werden? (Default TRUE)}
}
\value{
adaptiertes SVG als XML document
}
\description{
Passt ein in der SVG-Datei vorgefertigtes Differenzbalkendiagramm horizontal oder vertikal an. Vorbereitung: Balkensegmente und Wertelabels sind im SVG zu gruppieren. Mehrere solche Gruppen (Balken) können wiederum gruppiert werden. Die äußerste Gruppe ist zu benennen.
}
