% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{display_svg}
\alias{display_svg}
\title{Zeigt eingelesenes SVG an}
\usage{
display_svg(svg, width = NULL, height = NULL)
}
\arguments{
\item{svg}{SVG als XML document}

\item{width}{gewuenschte Breite (in Pixel) der Ausgabe (default: NULL)}

\item{height}{gewuenschte Hoehe (in Pixel) der Ausgabe (default: NULL)}
}
\description{
Die Funktionalitaet haengt von der Entwicklungsumgebung ab. In RStudio wird das SVG im Viewer angezeigt.
}
\details{
Wenn weder Breite noch Hoehe angegeben werden, wird die Originalgroesse (gegeben DPI) ausgegeben. Wenn nur einer dieser Parameter angegeben wird, wird der andere entsprechend des Originalverhaeltnisses automatisch skaliert.
}
