% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{linesSymbols_edit_circles}
\alias{linesSymbols_edit_circles}
\title{Lines and Symbols: Funktion fuer Kreise}
\usage{
linesSymbols_edit_circles(svg, group_name, frame_info, value_set,
  alignment, zeroLine = NULL)
}
\arguments{
\item{svg}{SVG als XML document}

\item{group_name}{Name der Gruppe mit den Kreiselementen (und optional Verbindungslinien).}

\item{frame_info}{Bereits eingelesene Informationen ueber den Grafiknamen (via frame_and_scaling eingelesen).}

\item{value_set}{Dataframe bzw. Vektor mit den Werten, eine Zeile pro Gruppe}

\item{alignment}{Ausrichtung der Symbole Entweder "horizontal" oder "vertical"}

\item{zeroLine}{Falls es sich um ein Differenzdiagramm mit 0er Linie handelt: Name der 0er Linie. (Default: NULL)}
}
\value{
adaptiertes SVG als XML document
}
\description{
Passt die Kreise (und Lininen) in der SVG-Datei vorgefertigtes Symboldiagramm horizontal oder vertikal an. Vorbereitung: Kreissegmente und Linien sind im SVG zu gruppieren. Die Gruppe ist zu benennen.
}
