% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{percentileBar}
\alias{percentileBar}
\title{Bar chart representing percentiles}
\usage{
percentileBar(
  svg,
  frame_name,
  group_name,
  scale_real,
  values,
  alignment = "horizontal",
  has_labels = TRUE,
  label_position = "center",
  decimals = 0,
  display_limits = c(0, 0)
)
}
\arguments{
\item{svg}{SVG als XML document}

\item{frame_name}{Name des Rechtseck-Elements mit dem Rahmen des Diagrammbereichs}

\item{group_name}{Name der Gruppe mit den Balkenelementen bzw. mehreren Balken}

\item{scale_real}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100))}

\item{values}{Vector or dataframe with percentile values}

\item{alignment}{Ausrichtung der Balken. Entweder "horizontal" (default) oder "vertical"}
}
\value{
adaptiertes SVG als XML document
}
\description{
Bar chart representing percentiles
}
