% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{read_svg}
\alias{read_svg}
\title{Liest eine SVG-Datei ein}
\usage{
read_svg(file, enc = "UTF-8", summary = FALSE, display = FALSE)
}
\arguments{
\item{file}{Dateiname (inkl. Pfad)}

\item{enc}{Encoding (default: "UTF-8")}

\item{summary}{Soll Zusammenfassung ueber SVG-File (Anzahl Gruppen, verwendete Farben...) angezeigt werden? (default: FALSE)}

\item{display}{Soll eingelesenes SVG (bspw. im R-Studio Viewer) angezeigt werden? (default: FALSE)}
}
\value{
SVG als XML document
}
\description{
Liest eine SVG-Datei ein
}
