% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{referenceBar}
\alias{referenceBar}
\title{Bar chart that is aligned around a reference category}
\usage{
referenceBar(
  svg,
  frame_name,
  group_name,
  scale_real,
  values,
  reference,
  nullvalue = 0,
  alignment = "horizontal",
  has_labels = TRUE,
  label_position = "center",
  decimals = 0,
  display_limits = 0
)
}
\arguments{
\item{svg}{SVG als XML document}

\item{frame_name}{Name des Rechtseck-Elements mit dem Rahmen des Diagrammbereichs}

\item{group_name}{Name der Gruppe mit den Balkenelementen bzw. mehreren Balken}

\item{scale_real}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100))}

\item{values}{Dataframe mit den Werten, eine Zeile pro Balken}

\item{reference}{Reference category (=column number of values). Bar segments up to this category lie to the left (horizontal) or to the bottom (vertical), bar segments above this category lie to the right (horizontal) or the top (vertical) of the bar chart.}

\item{nullvalue}{Value that defines the "center" of the bar segments (for left/right or bottom/top positioning)}

\item{alignment}{Ausrichtung der Balken. Entweder "horizontal" (default) oder "vertical"}

\item{has_labels}{Sollen Wertelabels angezeigt werden? (Default TRUE)}

\item{label_position}{Position der Wertelabels (Default "center")}

\item{decimals}{Anzahl der Dezimalstellen der Wertelabels (Default 0)}

\item{display_limits}{Interval for (small) values, that lead to suppression of the corresponding value labels. If only one value x is given, it is turned into the interval c(-x,x). (Default 0 = no suppression)}
}
\value{
adaptiertes SVG als XML document
}
\description{
Bar chart that is aligned around a reference category
}
