% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{scatterSymbols}
\alias{scatterSymbols}
\title{Passt Streudiagramme an}
\usage{
scatterSymbols(
  svg,
  frame_name,
  group_name,
  scale_real_x,
  scale_real_y,
  values,
  symbol_type
)
}
\arguments{
\item{svg}{SVG als XML document}

\item{frame_name}{Name des Grafikrahmens.}

\item{group_name}{Name der Gruppe mit der Symbolvorlage}

\item{scale_real_x}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100)) auf der x-Achse}

\item{scale_real_y}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100)) auf der y-Achse}

\item{values}{Dataframe bzw. Matrix mit den Werten, erste Spalte fuer X und zweite Spalte fuer Y}

\item{symbol_type}{Typ der Symbolgruppe, die angepasst werden soll: c("circle","rect","polygon","linegroup")}
}
\value{
adaptiertes SVG als XML document
}
\description{
Vervielfaeltigt und passt ein in der SVG-Datei vorgefertigtes Symbol horizontal und vertikal an. Vorbereitung: Einzelnes Symbol ist in SVG zu gruppieren. Die Gruppe ist zu benennen.
}
