% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{summary_svg}
\alias{summary_svg}
\title{Gibt eine Summary von eingelesenem SVG auf der Konsole aus}
\usage{
summary_svg(svg)
}
\arguments{
\item{svg}{SVG als XML document}
}
\description{
Gibt eine Summary von eingelesenem SVG auf der Konsole aus
}
