% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools-package.R
\docType{package}
\name{svgtools-package}
\alias{svgtools}
\alias{svgtools-package}
\title{svgtools: Manipulate SVG (Template) Files of Charts}
\description{
The purpose of this package is to manipulate SVG files that are templates of charts the user wants to produce. In vector graphics one copes with x-/y-coordinates of elements (e.g. lines, rectangles, text). Their scale is often dependent on the program that is used to produce the graphics. In applied statistics one usually has numeric values on a fixed scale (e.g. percentage values between 0 and 100) to show in a chart. Basically, 'svgtools' transforms the statistical values into coordinates and widths/heights of the vector graphics. This is done by stackedBar() for bar charts, by linesSymbols() for charts with lines and/or symbols (dot markers) and scatterSymbols() for scatterplots.
}
\details{
Package options:
\itemize{
 \item{\emph{svgtools.roundAwayFromZero}: When set to TRUE, use rounding away from zero (eg. 0.5 becomes 1) instead of \code{base::round()}; default is FALSE}
}
}
\author{
\strong{Maintainer}: Christian Wimmer \email{christian.wimmer@iqs.gv.at}

Authors:
\itemize{
  \item Konrad Oberwimmer
}

Other contributors:
\itemize{
  \item Michael Bruneforth [contributor]
}

}
\keyword{internal}
