% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{write_svg}
\alias{write_svg}
\title{Schreibt eingelesenes SVG in Datei}
\usage{
write_svg(svg, file)
}
\arguments{
\item{svg}{SVG als XML document}

\item{file}{Dateiname (inkl. Pfad)}
}
\description{
Schreibt eingelesenes SVG in Datei
}
