% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{linesSymbols}
\alias{linesSymbols}
\title{Lines and Symbols: Hauptfunktion}
\usage{
linesSymbols(svg, frame_name, group_name, scale_real, values,
  alignment = "horizontal", has_lines = TRUE, symbol_type = NULL,
  zero_line = NULL)
}
\arguments{
\item{svg}{SVG als XML document}

\item{frame_name}{Name des Grafikrahmens.}

\item{group_name}{Name der Gruppe mit den zu bearbeitenden Symbolen (und optional Verbindungslinien).}

\item{scale_real}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100))}

\item{values}{Dataframe bzw. Vektor mit den Werten, eine Zeile pro Gruppe}

\item{alignment}{Ausrichtung der Symbole. Entweder "horizontal" oder "vertical" (Default: horizontal).}

\item{has_lines}{Sind Linien vorhanden? (Default = TRUE)}

\item{symbol_type}{Typ der Symbolgruppe, die angepasst werden soll. Derzeit moeglich: "circle" und "rect" (Default: NULL = nur Linie).}

\item{zero_line}{Falls es sich um ein Differenzdiagramm mit 0er Linie handelt: Name der 0er Linie. (Default: NULL)}
}
\value{
adaptiertes SVG als XML document
}
\description{
Passt die Symbole (und Lininen) in der SVG-Datei vorgefertigtes Symboldiagramm horizontal oder vertikal an. Vorbereitung: Kreissegmente und Linien sind im SVG zu gruppieren. Die Gruppe ist zu benennen.
}
