% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgtools.R
\name{stackedBar}
\alias{stackedBar}
\title{Passt Balkendiagramm an}
\usage{
stackedBar(svg, frame_name, group_name, scale_real, values,
  alignment = "horizontal", has_labels = TRUE, label_position = "center",
  decimals = 0, display_limit = 0)
}
\arguments{
\item{svg}{SVG als XML document}

\item{frame_name}{Name des Rechtseck-Elements mit dem Rahmen des Diagrammbereichs}

\item{group_name}{Name der Gruppe mit den Balkenelementen bzw. mehreren Balken}

\item{scale_real}{Unter- und Obergrenze des dargestellten Wertebereichs (bspw. c(0,100))}

\item{values}{Dataframe mit den Werten, eine Zeile pro Balken}

\item{alignment}{Ausrichtung der Balken. Entweder "horizontal" (default) oder "vertikal"}

\item{has_labels}{Sollen Wertelabels angezeigt werden? (Default TRUE)}

\item{label_position}{Position der Wertelabels (Default "center")}

\item{decimals}{Anzahl der Dezimalstellen der Wertelabels (Default 0)}

\item{display_limit}{Unter welchem Wert sollen Wertelabels unterdrückt werden? (Default 0)}
}
\value{
adaptiertes SVG als XML document
}
\description{
Passt ein in der SVG-Datei vorgefertigtes (gestapeltes) Balkendiagramm horizontal oder vertikal an. Vorbereitung: Balkensegmente und Wertelabels sind im SVG zu gruppieren. Mehrere solche Gruppen (Balken) können wiederum gruppiert werden. Die äußerste Gruppe ist zu benennen.
}
